Fix the KDF
Wasn't thinking before. It was a ratchet but without separating the output from the chain key. I believe this is why Argon2 truncates hashes in their variable-length hash function - you can't compute the next block from the previous output because half of the input is missing.

This change matches the traditional KDF in the HKDF paper, except it uses a feedback mode and Extract-then-Expand more like HKDF. Would be better to pad the key with zeros to the block size to induce a random IV but avoiding that for simplicity.