Improve the NIST KDF
Using the XOR of the outputs as the IV makes more sense than using it as the Context parameter. This is more efficient and avoids a suffix MAC situation.

However, with SHAKE256, SHA3-512, and BLAKE3, the current PRF input still exceeds one block. I think keeping the counter and domain separation is a good idea though so will probably have to live with this.