Add a mandatory personalization parameter
This is a great idea besides the issue of it potentially affecting performance when deriving the pseudorandom bytes. That's why I've restricted the length, although MAX_PERSONALIZATION would cause the input to exceed one block with SHA-224/SHA-256. It may need to be increased, although the BLAKE2b personalization is only 16 bytes.