Allow a longer personalization parameter
If you cache the hash function state, it doesn't matter that this causes the input to go into more than one block. However, to avoid this caching, I could change the pseudorandom bytes derivation to hash everything but the counter and use that as the PRF key. Then the counter would be the only thing in the PRF message.